Refactor: collapse Dialog+Form pairs into single modules per operation
Each task operation (create / update / remove) was split across a thin
Dialog wrapper, a long Form, and a button — three files to read for one
concept, with the Dialog adding nothing beyond `<Dialog open={…}>`.
The Form already called close() itself, so the wrapper was earning no
locality.

- Inline the Dialog wrapping into the Form file
- Rename CreateTaskForm → CreateTask (and similarly for Update / Remove)
  to reflect that the file now owns the whole operation
- Delete the three Dialog files
- Update ManageTasks.tsx imports
