--- src/modules/main/Pages/to-do/components/ManageTasks/Components/remove-task/RemoveTaskForm.tsx
@@ -1,107 +1,0 @@
-import { Grid, Paper, Typography, useTheme } from '@mui/material';
-import { Form, Formik } from 'formik';
-import ButtonWrapper from '../../../../../../../global/components/ButtonWrapper';
-import TextfieldWrapper from '../../../../../../../global/components/TextfieldWrapper';
-import { tokens } from '../../../../../../../global/theme/theme';
-import useDialogStore from '../../../../hooks/useDialogStore';
-import { useTaskForm } from '../../hooks/useTaskForm';
-
-function RemoveTaskForm() {
-  const theme = useTheme();
-  const colors = tokens(theme.palette.mode);
-  const {
-    DELETE_FORM_VALIDATION,
-    INITIAL_REMOVE_FORM_STATE,
-    handleDeleteTask,
-    buttonConfig,
-  } = useTaskForm();
-  const close = useDialogStore((s) => s.close);
-
-  return (
-    <Grid
-      container
-      display="flex"
-      justifyContent="center"
-      alignItems="center"
-      height="90%"
-    >
-      <Paper
-        sx={{
-          padding: 4,
-          display: 'flex',
-          flexDirection: 'row',
-          justifyContent: 'center',
-          backgroundColor: `${colors.primary[400]}`,
-        }}
-        variant="outlined"
-      >
-        <Formik
-          validationSchema={DELETE_FORM_VALIDATION}
-          initialValues={{ ...INITIAL_REMOVE_FORM_STATE }}
-          onSubmit={handleDeleteTask}
-        >
-          {({ submitForm, isValid }) => (
-            <Form style={{ width: '60%' }}>
-              <Grid container spacing={3}>
-                <Grid item xs={12}>
-                  <Typography fontWeight={'bold'}>
-                    Enter the Title of the Task to Delete
-                  </Typography>
-                </Grid>
-                <Grid item xs={12}>
-                  <TextfieldWrapper
-                    name="taskTitle"
-                    label="Title"
-                    type="text"
-                  />
-                </Grid>
-                <Grid item xs={12}>
-                  <ButtonWrapper
-                    buttonConfig={{
-                      ...buttonConfig,
-                      sx: {
-                        bgcolor: '#eb3434',
-                        '&:hover': {
-                          backgroundColor: '#d62d2d',
-                        },
-                      },
-                      onClick: () => {
-                        if (isValid) {
-                          submitForm();
-                          close();
-                        }
-                      },
-                    }}
-                  >
-                    Delete Task and Exit
-                  </ButtonWrapper>
-                  <ButtonWrapper
-                    buttonConfig={{
-                      ...buttonConfig,
-                      sx: {
-                        bgcolor: '#cc0000',
-                        '&:hover': {
-                          backgroundColor: '#b30000',
-                        },
-                        marginTop: 2,
-                      },
-                      onClick: () => {
-                        if (isValid) {
-                          submitForm();
-                        }
-                      },
-                    }}
-                  >
-                    Delete Task and Continue
-                  </ButtonWrapper>
-                </Grid>
-              </Grid>
-            </Form>
-          )}
-        </Formik>
-      </Paper>
-    </Grid>
-  );
-}
-
-export default RemoveTaskForm;
